Refactored Stream/SubscriptionResponder/Requester
