+++ src/automata/StreamRequester.h
@@ -9,11 +9,11 @@
 #include "src/Frame.h"
 #include "src/Payload.h"
 #include "src/ReactiveStreamsCompat.h"
+#include "src/automata/StreamSubscriptionRequesterBase.h"
 #include "src/mixins/ConsumerMixin.h"
 #include "src/mixins/ExecutorMixin.h"
 #include "src/mixins/LoggingMixin.h"
 #include "src/mixins/MemoryMixin.h"
-#include "src/mixins/MixinTerminator.h"
 #include "src/mixins/SourceIfMixin.h"
 #include "src/mixins/StreamIfMixin.h"
 
@@ -27,50 +27,18 @@ enum class StreamCompletionSignal;
 
 /// Implementation of stream automaton that represents a Stream requester
 class StreamRequesterBase
-    : public LoggingMixin<ConsumerMixin<Frame_RESPONSE, MixinTerminator>> {
-  using Base = LoggingMixin<ConsumerMixin<Frame_RESPONSE, MixinTerminator>>;
+    : public StreamSubscriptionRequesterBase<Frame_REQUEST_STREAM> {
+  using Base = StreamSubscriptionRequesterBase<Frame_REQUEST_STREAM>;
 
  public:
   using Base::Base;
 
-  /// Degenerator form of the Subscriber interface -- only one request payload
-  /// will be sent to the server.
-  void onNext(Payload);
-
-  /// @{
-  /// A Subscriber interface to control ingestion of response payloads.
-  void request(size_t);
-
-  void cancel();
-  /// @}
-
  protected:
   /// @{
-  void endStream(StreamCompletionSignal);
-
-  /// Not all frames are intercepted, some just pass through
-  using Base::onNextFrame;
-
-  void onNextFrame(Frame_RESPONSE&);
-
-  void onNextFrame(Frame_ERROR&);
-
   std::ostream& logPrefix(std::ostream& os);
   /// @}
-
- private:
-  /// State of the Stream requester.
-  enum class State : uint8_t {
-    NEW,
-    REQUESTED,
-    CLOSED,
-  } state_{State::NEW};
-  /// An allowance accumulated before the stream is initialized.
-  /// Remaining part of the allowance is forwarded to the ConsumerMixin
-  reactivestreams::AllowanceSemaphore initialResponseAllowance_;
 };
 
-using StreamRequester =
-    SourceIfMixin<StreamIfMixin<LoggingMixin<ExecutorMixin<
-        LoggingMixin<MemoryMixin<LoggingMixin<StreamRequesterBase>>>>>>>;
+using StreamRequester = SourceIfMixin<StreamIfMixin<LoggingMixin<ExecutorMixin<
+    LoggingMixin<MemoryMixin<LoggingMixin<StreamRequesterBase>>>>>>>;
 }
