+++ src/automata/StreamResponder.cpp
@@ -13,61 +13,6 @@
 
 namespace reactivesocket {
 
-void StreamResponderBase::onNext(Payload response) {
-  switch (state_) {
-    case State::RESPONDING:
-      Base::onNext(std::move(response));
-      break;
-    case State::CLOSED:
-      break;
-  }
-}
-
-void StreamResponderBase::onComplete() {
-  switch (state_) {
-    case State::RESPONDING: {
-      state_ = State::CLOSED;
-      Frame_RESPONSE frame(
-          streamId_, FrameFlags_COMPLETE, FrameMetadata::empty(), nullptr);
-      connection_->onNextFrame(frame);
-      connection_->endStream(streamId_, StreamCompletionSignal::GRACEFUL);
-    } break;
-    case State::CLOSED:
-      break;
-  }
-}
-
-void StreamResponderBase::onError(folly::exception_wrapper ex) {
-  switch (state_) {
-    case State::RESPONDING: {
-      state_ = State::CLOSED;
-      Frame_ERROR frame(
-          streamId_,
-          FrameFlags_EMPTY,
-          ErrorCode::APPLICATION_ERROR,
-          FrameMetadata::empty(),
-          folly::IOBuf::copyBuffer(ex.what().toStdString()));
-      connection_->onNextFrame(frame);
-      connection_->endStream(streamId_, StreamCompletionSignal::GRACEFUL);
-    } break;
-    case State::CLOSED:
-      break;
-  }
-}
-
-void StreamResponderBase::endStream(StreamCompletionSignal signal) {
-  switch (state_) {
-    case State::RESPONDING:
-      // Spontaneous ::endStream signal means an error.
-      DCHECK(StreamCompletionSignal::GRACEFUL != signal);
-      state_ = State::CLOSED;
-      break;
-    case State::CLOSED:
-      break;
-  }
-  Base::endStream(signal);
-}
-
 void StreamResponderBase::onNextFrame(Frame_REQUEST_STREAM& frame) {
   bool end = false;
   switch (state_) {
@@ -86,19 +31,7 @@ void StreamResponderBase::onNextFrame(Frame_REQUEST_STREAM& frame) {
   }
 }
 
-void StreamResponderBase::onNextFrame(Frame_CANCEL& frame) {
-  switch (state_) {
-    case State::RESPONDING:
-      state_ = State::CLOSED;
-      connection_->endStream(streamId_, StreamCompletionSignal::GRACEFUL);
-      break;
-    case State::CLOSED:
-      break;
-  }
-}
-
 std::ostream& StreamResponderBase::logPrefix(std::ostream& os) {
-  return os << "StreamResponder(" << &connection_ << ", " << streamId_
-            << "): ";
+  return os << "StreamResponder(" << &connection_ << ", " << streamId_ << "): ";
 }
 }
