+++ src/automata/StreamSubscriptionResponderBase.h
@@ -10,10 +10,11 @@
 #include "src/Frame.h"
 #include "src/Payload.h"
 #include "src/ReactiveStreamsCompat.h"
-#include "src/automata/StreamSubscriptionResponderBase.h"
 #include "src/mixins/ExecutorMixin.h"
 #include "src/mixins/LoggingMixin.h"
 #include "src/mixins/MemoryMixin.h"
+#include "src/mixins/MixinTerminator.h"
+#include "src/mixins/PublisherMixin.h"
 #include "src/mixins/SinkIfMixin.h"
 #include "src/mixins/StreamIfMixin.h"
 
@@ -25,25 +26,41 @@ namespace reactivesocket {
 
 enum class StreamCompletionSignal;
 
-/// Implementation of stream automaton that represents a Stream responder
-class StreamResponderBase : public StreamSubscriptionResponderBase {
-  using Base = StreamSubscriptionResponderBase;
+/// Implementation of stream automaton that represents a Stream/Subscription
+/// responder.
+class StreamSubscriptionResponderBase
+    : public LoggingMixin<PublisherMixin<Frame_RESPONSE, MixinTerminator>> {
+  using Base = LoggingMixin<PublisherMixin<Frame_RESPONSE, MixinTerminator>>;
 
  public:
   using Base::Base;
 
+  /// @{
+  /// A Subscriber implementation exposed to the user of ReactiveSocket to
+  /// receive "response" payloads.
+  void onNext(Payload);
+
+  void onComplete();
+
+  void onError(folly::exception_wrapper);
+  /// @}
+
  protected:
   /// @{
+  void endStream(StreamCompletionSignal);
 
   /// Not all frames are intercepted, some just pass through.
   using Base::onNextFrame;
 
-  void onNextFrame(Frame_REQUEST_STREAM&);
+  void onNextFrame(Frame_CANCEL&);
 
   std::ostream& logPrefix(std::ostream& os);
   /// @}
-};
 
-using StreamResponder = SinkIfMixin<StreamIfMixin<LoggingMixin<ExecutorMixin<
-    LoggingMixin<MemoryMixin<LoggingMixin<StreamResponderBase>>>>>>>;
+  /// State of the Subscription responder.
+  enum class State : uint8_t {
+    RESPONDING,
+    CLOSED,
+  } state_{State::RESPONDING};
+};
 }
