+++ src/automata/SubscriptionResponder.h
@@ -10,11 +10,10 @@
 #include "src/Frame.h"
 #include "src/Payload.h"
 #include "src/ReactiveStreamsCompat.h"
+#include "src/automata/StreamSubscriptionResponderBase.h"
 #include "src/mixins/ExecutorMixin.h"
 #include "src/mixins/LoggingMixin.h"
 #include "src/mixins/MemoryMixin.h"
-#include "src/mixins/MixinTerminator.h"
-#include "src/mixins/PublisherMixin.h"
 #include "src/mixins/SinkIfMixin.h"
 #include "src/mixins/StreamIfMixin.h"
 
@@ -27,43 +26,21 @@ namespace reactivesocket {
 enum class StreamCompletionSignal;
 
 /// Implementation of stream automaton that represents a Subscription responder.
-class SubscriptionResponderBase
-    : public LoggingMixin<PublisherMixin<Frame_RESPONSE, MixinTerminator>> {
-  using Base = LoggingMixin<PublisherMixin<Frame_RESPONSE, MixinTerminator>>;
+class SubscriptionResponderBase : public StreamSubscriptionResponderBase {
+  using Base = StreamSubscriptionResponderBase;
 
  public:
   using Base::Base;
 
-  /// @{
-  /// A Subscriber implementation exposed to the user of ReactiveSocket to
-  /// receive "response" payloads.
-  void onNext(Payload);
-
-  void onComplete();
-
-  void onError(folly::exception_wrapper);
-  /// @}
-
  protected:
   /// @{
-  void endStream(StreamCompletionSignal);
-
   /// Not all frames are intercepted, some just pass through.
   using Base::onNextFrame;
 
   void onNextFrame(Frame_REQUEST_SUB&);
 
-  void onNextFrame(Frame_CANCEL&);
-
   std::ostream& logPrefix(std::ostream& os);
   /// @}
-
- private:
-  /// State of the Subscription responder.
-  enum class State : uint8_t {
-    RESPONDING,
-    CLOSED,
-  } state_{State::RESPONDING};
 };
 
 using SubscriptionResponder =
